backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-docx==0.8.11
reportlab==4.0.4
pandas==2.1.3
python-multipart==0.0.6
python-docx-template==0.16.9
jinja2==3.1.2
psycopg2-binary==2.9.9